Continuing work of https://github.com/kubernetes/kubernetes/pull/108392

Started from taking the exact implementation by @MadhavJivrajani, however updated the btree factor to 8. 

Using btree without any changes already is faster than existing ThreadedSafeStore. There are still many improvements that could be implemented. Will continue on it when I have time.

/kind feature

```release-note
NONE
```
                       │    old.txt    │               new.txt               │
                       │    sec/op     │    sec/op     vs base               │
StoreCreateGet-24         576.6µ ± 15%   635.8µ ±  7%        ~ (p=0.132 n=6)
StoreList-24              770.2µ ±  5%   708.6µ ± 43%        ~ (p=0.132 n=6)
StoreListPaginate-24     18.302m ±  6%   8.938m ±  3%  -51.16% (p=0.002 n=6)
StoreListNodeScoped-24    444.7µ ± 12%   279.1µ ±  3%  -37.25% (p=0.002 n=6)
StoreListNamespace-24     6.077m ± 10%   1.493m ±  2%  -75.44% (p=0.002 n=6)
geomean                   1.855m         1.109m        -40.22%
```