@serathius 2024-08-22
Addressed all the comments, will rerun benchmarks to see if this impacted results.

PTAL @wojtek-t