@serathius 2024-08-22
Need to double check but migrating to apiserver utils rand made the benchmark much slower