@liggitt 2024-08-26
I didn't review the test additions closely, will lean on wojtek's review. One question about the unconditional typecast in the test indexer, then lgtm